% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api.R
\name{.fetch_siconfi_data}
\alias{.fetch_siconfi_data}
\title{Request Template para SICONFI}
\usage{
.fetch_siconfi_data(url)
}
\arguments{
\item{url}{String. Endpoint da API.}
}
\value{
Objeto httr.
}
\description{
Abstração para chamadas à API com tratamento de erros.
}
