% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGFdata.R
\name{RGFdata}
\alias{RGFdata}
\title{RGFdata: Extração intuitiva e fácil dos dados do RGF}
\usage{
RGFdata(
  cod.ibge = NULL,
  ano = NULL,
  poder = NULL,
  periodo = NULL,
  anexo = NULL,
  simplificado = FALSE
)
}
\arguments{
\item{cod.ibge}{Código IBGE do Ente. se \code{simplificado = TRUE}, então selecione municípios.}

\item{ano}{Exercício do relatório}

\item{poder}{Código do poder. Valores disponíveis: E = Executivo, L = Legislativo, J = Judiciário, M = Ministério Público, D = Defensoria Pública}

\item{periodo}{Quadrimestre ou semestre de referência do relatório dentro de um exercício. A periodicidade semestral é automaticamente selecionada se \code{simplificado = TRUE}.  A periodicidade padrão do relatório é quadrimestral. Valores disponíveis: 1, 2 ou 3.}

\item{anexo}{Anexos dos demonstrativos do RGF. Valores disponíveis: 1, 2, 3, 4, 5 ou 6.}

\item{simplificado}{Tipo do Demonstrativo. RGF Simplificado aplica-se apenas aos municípios com menos de 50 mil habitantes que optaram pela publicação semestral dos relatórios. Se \code{TRUE} a periodicidade semestral será automaticamente selecionada.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Realiza a extração dos dados do Relatório de Gestão Fiscal (RGF) de maneira mais intuitiva e fácil utilizando a API do \href{https://apidatalake.tesouro.gov.br/docs/siconfi/}{SICONFI}.
}
\examples{
# install.packages("devtools")
# devtools::install_github("Natanaelsl/RREORGFdataR")

# Carregando o pacote
# library(RREORGFdataR)

# Extraindo dados do anexo 1 para o 3º quadrimestre
# do RGF de 2020 até 2023 do Estado de Goiás para todos os poderes.
# RGFdata(cod.ibge = 52,
#         ano = c(2020:2023),
#         poder = c('E','L','J','D','M'),
#         periodo = 3,
#         anexo = 1,
#         simplificado = FALSE)
}
