% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RREORGFdataR-package.R
\docType{package}
\name{RREORGFdataR-package}
\alias{RREORGFdataR}
\alias{RREORGFdataR-package}
\title{RREORGFdataR: Extração intuitiva e fácil dos dados do RREO e RGF utilizando API do SICONFI.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='Logo_1' width='120'}}
O 'RREORGFdataR' é um pacote R que permite aos usuários acessar facilmente o conjunto de dados do Relatório Resumido da Execução Orçamentária (RREO) e do Relatório de Gestão Fiscal (RGF) utilizando a API do SICONFI (https://apidatalake.tesouro.gov.br/docs/siconfi/). Saiba mais sobre a utilização do 'RREORGFdataR' em \url{https://natanaelsl.com.br/project/rreorgf_pkg/} ou \url{https://natanaelsl.github.io/RREORGFdataR/}.
}
\details{
O Tesouro Nacional disponibilizou a Application Programming Interface (API) de dados abertos para atender à demanda por dados brutos oriundos do Sistema de Informações Contábeis e Fiscais do Setor Público Brasileiro – Siconfi. Por meio dessa ferramenta, é possível ao usuário obter desde pequenas frações até grandes volumes de dados de todas as informações inseridas pelos entes subnacionais no Siconfi.
}
\seealso{
Useful links:
\itemize{
\item \url{https://natanaelsl.com.br/project/rreorgf_pkg/}
\item \url{https://natanaelsl.github.io/RREORGFdataR/}
\item \url{https://github.com/Natanaelsl/RREORGFdataR}
\item Report bugs at \url{https://github.com/Natanaelsl/RREORGFdataR/issues}
}
}
\author{
\itemize{
\item Natanael Soares Leite (\href{https://orcid.org/0000-0003-1845-7095}{ORCID})
\item Raphael Maciel de Lima
}
}
\keyword{internal}
