% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RREOdata.R
\name{RREOdata}
\alias{RREOdata}
\title{RGFdata: Extração intuitiva e fácil dos dados do RREO}
\usage{
RREOdata(
  cod.ibge = NULL,
  year = NULL,
  period = NULL,
  annex = NULL,
  simplified = FALSE
)
}
\arguments{
\item{cod.ibge}{Código IBGE do Ente. se \code{simplified = TRUE}, então selecione municípios.}

\item{year}{Exercício do relatório}

\item{period}{Quadrimestre ou semestre de referência do relatório dentro de um exercício. A periodicidade semestral é automaticamente selecionada se \code{simplified = TRUE}.  A periodicidade padrão do relatório é quadrimestral. Valores disponíveis: 1, 2 ou 3.}

\item{annex}{Anexos dos demonstrativos do RREO. Valores disponíveis: 1, 2, 3, 4, 5 ou 6.}

\item{simplified}{Tipo do Demonstrativo. RREO Simplificado aplica-se apenas aos municípios com menos de 50 mil habitantes que optaram pela publicação semestral dos relatórios. Se \code{TRUE} a periodicidade semestral será automaticamente selecionada.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Realiza a extração dos dados do Relatório Resumido de Execução Orçamentária (RREO) de maneira mais intuitiva e fácil utilizando a API do \href{https://apidatalake.tesouro.gov.br/docs/siconfi/}{SICONFI}.
}
\examples{
# install.packages("devtools")
# devtools::install_github("Natanaelsl/RREORGFdataR")

# Carregando o pacote
library(RREORGFdataR)

# Extraindo dados do anexo 1, 6º bimestre
# do RREO de 2020 até 2023 do Estado de Goiás.
df.RREO <- RREOdata(cod.ibge = 52,
                    year = c(2020:2023),
                    period = 6,
                    annex = 1,
                    simplified = FALSE)

}
