% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siconfi_list.R
\name{siconfi_list}
\alias{siconfi_list}
\title{siconfi_list: Lista de instituições e respectivos códigos}
\usage{
siconfi_list(options = NULL)
}
\arguments{
\item{options}{A opção padrão \code{NULL} retorna um \code{data.frame} contendo os códigos do IBGE e do Siconfi disponíveis para a API, juntamente com as instituições correspondentes. A escolha da opção \code{options = down} permite o download do arquivo (.pdf) disponibilizado pelo Tesouro Nacional.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Retorna informações sobre os códigos do IBGE e do Siconfi e as respectivas instituições compatíveis com a API do SICONFI.
}
\examples{
# install.packages("devtools")
devtools::install_github("Natanaelsl/RREORGFdataR")

# Carregando o pacote
library(RREORGFdataR)

# Reportando `data.frame` com os códigos disponibilizados pelo Tesouro Nacional
siconfi_list() \%>\%
utils::tail(10)

# Baixando arquivo (.pdf) disponibilizado pelo Tesouro Nacional
# apresentando lista com os códigos
# siconfi_list(options == "down")
}
